Feat: Added View Course component
Merged Pull Request
