+++ src/App.css
@@ -13,8 +13,7 @@
   }
 }
 
-
-.toast-header .btn-close{
+.toast-header .btn-close {
   display: none;
 }
