+++ src/AppRouter.js
@@ -1,24 +1,23 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import { HashRouter as Router } from 'react-router-dom';
+import { HashRouter as Router } from "react-router-dom";
 import { createBrowserHistory } from "history";
 import App from "./App.js";
 import Special from "./Special.js";
-import Courses from "./Courses.js";
+import Courses from "./Courses.jsx";
 import Dashboard from "./Dashboard.js";
 
 const history = createBrowserHistory();
 
 function AppRouter() {
-
   return (
-    <Router basename='/' history={history}>
-        <Routes>
-          <Route path="/specials" element={<Special/>} />
-          <Route path="/courses" element={<Courses/>} />
-          <Route path="/dashboard" element={<Dashboard/>} />
-          <Route path="/" element={<App/>} />
-        </Routes>
+    <Router basename="/" history={history}>
+      <Routes>
+        <Route path="/specials" element={<Special />} />
+        <Route path="/courses" element={<Courses />} />
+        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/" element={<App />} />
+      </Routes>
     </Router>
   );
 }
