+++ src/index.css
@@ -1,3 +1,7 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
@@ -16,7 +20,3 @@ html,
 body {
   background-color: #000;
 }
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
